arch: break up main.rs — extract CLI, commands, serve (#143)
Extract 2031-line main.rs into focused modules:
- cli.rs: clap structs and argument parsing (339 lines)
- serve.rs: serve() orchestration + query_live_agents() (215 lines)
- commands/: one module per subcommand group (agent, graph, sm, task,
  schedule, status, restart, upgrade, remind) + shared helpers
- main.rs: thin entry point with dispatch (~95 lines + tests)
